requirements.txt
Django 
djangorestframework 
google-generativeai 
pillow 
python-decouple